Use separate factories for each build step
